traps/dart: merge DartEmitter to dart.c
