+++ src/game/traps/dart.h
@@ -7,3 +7,6 @@ void Dart_Control(int16_t item_num);
 
 void DartEffect_Setup(OBJECT_INFO *obj);
 void DartEffect_Control(int16_t fx_num);
+
+void DartEmitter_Setup(OBJECT_INFO *obj);
+void DartEmitter_Control(int16_t item_num);
