--- src/game/traps/dart_emitter.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-void DartEmitter_Setup(OBJECT_INFO *obj);
-void DartEmitter_Control(int16_t item_num);
